Support ECMA 262 regular expression anchors
Ruby uses `^` in regular expressions to match the beginning of a line
(even in multiline strings) while in ECMA 262 it only matches the
beginning of the entire string. Same for `$` and the end of a line. This
parses the provided regular expressions and replaces the ECMA 262
anchors with their Ruby equivalents (`\A` and `\z`).

Fixes #11