Remove regexp_parser dependency
json_schemer doesn't seem to directly require or depend on
regexp_parser. It requires ecma-re-validator, which transitively
dependes on regexp_parser. So we should only require on
ecma-re-validator, and allow it to require its own upstream
dependencies.

The current dependency is also too strict and is preventing
regexp_parser 1.6.0 being used in other gems.