Format pretty messages from error hash
This is based on a snippet of code offered by @robacarp in #48. The code has
been cleaned up and tested and provides a standalone interface for converting
errors into a human readable message:

```ruby
errors = schema.validate({ 'a' => { 'x' => 1 } }).to_a
errors.map {|err| JSONSchemer::Errors.pretty err }
```

Longer term, an option can be introduced that allows the generation of the
message inline with the validation. For example, given the option `pretty:
true`, then on line 229 of `lib/json_schemer/schema/base.rb` (as of #a2a6df8),
the message could be added to the error hash:

```ruby
  error['message'] = JSONSchemer::Errors.pretty(error) if pretty
```