Fixes for ruby 3.0.0.preview1
I replaced the call to URI.escape with the contents of URI.escape
such as it was, because it's removed in ruby 3.0.0.preview1
I assume there's a reason it's removed, so this is probably incomplete
at best, but at least it should continue to work as it did.

For this reason, & because 'uri_template' gem was throwing up a lot of
errors it may be a good idea to move to using the 'addressable' gem. but
for now i wanted to make as few changes as possible.

The only other change was adding an explicit splat to kwargs thing, and
changing the required ruby version to match 3.0.0.preview1 also.