Stop rewriting `\A` to literal `A` in regexes
I introduced this behavior in: https://github.com/davishmcclurg/json_schemer/pull/12
but thinking about it again, it seems more annoying than it is helpful.
If someone writes `\A` in their regex, they're probably writing it in
Ruby with the expectation that it's beginning-of-string. It seems less
likely that they're writing an ECMA 262 regex with the expectation that
it matches a literal `A` (because why would they write it `\A`?).