Add CLI
This adds a simple script to validate data from the command line. It
takes a JSON schema file as the first argument followed by one or more
JSON data files to validate. If there are any validation errors, it
outputs them and returns an error code.

Validation errors are output as single-line JSON objects. The `--errors`
option can be used to limit the number of errors returned or prevent
output entirely (and fail fast).

The schema or data can also be read from stdin using `-`.

Addresses: https://github.com/davishmcclurg/json_schemer/issues/95

I used `ARGF` to make reading from files/stdin easier, but had to
convert the schema file to `Pathname` to get refs to work intuitively.

It was tricky to test. To get it working on all platforms, I have it
building the gem to a temp directory that I add to `GEM_PATH` when
running the executable in tests. I also had to override `RUBYOPT` to
prevent Bundler for messing with things. I think there's some benefit to
building the gem in CI as well.