Escape file paths when generating URI
File paths with spaces are causing `URI::InvalidURIError` errors:

```
>> URI.parse(File.join('file:', '/path/t o/schema.json'))
/Users/dharsha/.rbenv/versions/3.1.2/lib/ruby/3.1.0/uri/rfc3986_parser.rb:67:in `split': bad URI(is not URI?): "file:/path/t o/schema.json" (URI::InvalidURIError)
```

I think the correct fix is to escape paths when generating URIs and
unescape when reading files. I had a hard time finding the right
escape/unescape methods, though, since I think we want `%20` for spaces
instead of `+` and to leave slashes unescaped. `URI.escape` previously
did that but it's been deprecated for a while and removed in Ruby 3.
`URI::DEFAULT_PARSER` still has the same escape method and it doesn't
emit a warning, so I'm going with it.

Fixes: https://github.com/davishmcclurg/json_schemer/issues/104