Add named regexp resolvers
`ecma` maintains the current behavior of rewriting Ruby's `\A` and `\z`
anchors.
`ruby` uses Ruby's `Regexp` directly.

In the future, I think it may make sense to switch the default to `ruby`
since it's the least surprising behavior. That way people could opt in
to expression rewrites and the rewrites could be more aggressive.