Support regexp_parser < 2.3.0
In 2.3.0, regexp_parser switched to using `Syntax` classes directly
instead of creating instances. You get a bunch of warnings if you use
instances in later versions. To support earlier versions, this checks
where `implements` is defined and creates an instance if necessary.

Also had to stop using `Backreference::Plain` because it was introduced
in 2.2.0. This replaces it with its value: `%i[number]`.

I also switched to removing unsupported tokens from predefined sets so
that it's a little easier to understand how these were built:

- `(Escape::Control - %i[meta_sequence])`
- `(Escape::ASCII - %i[bell escape])`
- `(FreeSpace::All - %i[comment])`