Replace uri_template gem with regular expression
`uri_template` emits some annoying warnings for circular requires and it
doesn't seem like they're ever going to fix it. We don't need the
expansion functionality so I just replaced it with a regular expression
based off the [syntax description][0] in the rfc. I pulled all the test
examples from the rfc as well and added some edge cases and invalid
examples.

[0]: https://datatracker.ietf.org/doc/html/rfc6570#section-2