Return parent base URI when resolving ids
This is necessary because `$id` values are joined with the base URI when
the [instance is validated][0] as well as when [ref JSON pointers are
resolved][1]. This returns the parent base URI instead so that calling
`join_uri` later will generate the correct base URI. There's probably a
better way to do this but I get confused thinking it through.

This came up while working on adding support for draft 2019-09, which
has [tests][2] that caused issues. That's where I pulled these tests
from.

[0]: https://github.com/davishmcclurg/json_schemer/blob/c33c1b91ca831c01270cd7d14f927b8727022dee/lib/json_schemer/schema/base.rb#L137
[1]: https://github.com/davishmcclurg/json_schemer/blob/c33c1b91ca831c01270cd7d14f927b8727022dee/lib/json_schemer/schema/base.rb#L340
[2]: https://github.com/json-schema-org/JSON-Schema-Test-Suite/blob/b069ac352c3dc1fae71f6f177dafe97001c97920/tests/draft2019-09/refRemote.json#L101-L160