Resolve JSON pointer refs locally
When `instance.base_uri` is relaive and `ref` is a JSON pointer
fragment, the fragment is removed from `ref_uri` which causes it to be
empty. The ref is meant to be resolved locally, but it was being passed
to `resolve_ref` instead. I thought this case was covered by the
`ref_uri.to_s == @base_uri.to_s` check, but it's currently possible
(though incorrect?) for `@base_uri` to be relative.

Checking `ref_uri.to_s.empty?` should be ok since it means `ref` was a
plain JSON pointer fragment (eg, `#/path/to/schema`).

Example here: https://github.com/davishmcclurg/json_schemer/issues/131

Closes: https://github.com/davishmcclurg/json_schemer/issues/131