Better email validation
The existing regex doesn't support spaces in quotes and it seems better
to match the rfc as much as possible. I built the regex using `Mailbox`
in [rfc5321][0] and then added non-ascii support as described in
[rfc6531][1] for `idn-email`. Regular `email` uses the same check with
an `ascii_only?` guard to filter to ascii.

For `Domain` and `address-literal`, I'm using the existing checks
because writing those as regular expressions seems hard and I believe
the behavior matches the specification.

I pulled the test cases from json-schema-test-suite for draft 2020-12,
which is how I caught the problem.

[0]: https://datatracker.ietf.org/doc/html/rfc5321#section-4.1.2
[1]: https://datatracker.ietf.org/doc/html/rfc6531#section-3.3