`insert_property_defaults` in conditional schemas
This uses the validation `Result` tree to support inserting property
defaults in conditional subschemas. It looks through the results for
"properties" keywords that contain defaults, groups them by instance and
property, tracks if they're in a valid subtree, and passes them to the
resolver to do the actual insertion. The default resolver inserts the
default value if there's only one available or if there's only one
unique one from a valid subtree. If any defaults are inserted, the
validation is run again and that result is returned.

I got the test examples from the issues below. I think the default
behavior should make sense in most cases. Otherwise, a custom
`property_default_resolver` can be provided. The only special case is
that results from `not` subschemas are dropped. It might make more sense
to handle that in the resolver, but for now I think it's ok.

Related:

- https://github.com/davishmcclurg/json_schemer/issues/53
- https://github.com/davishmcclurg/json_schemer/issues/94