Short-circuit `valid?` on first keyword error
The `flag` output format (which `valid?` uses) doesn't return nested
errors or annotations, so it should be safe to short-circuit after the
first error. In the future, this could be extended to applicators
keywords like `allOf`, but for now I think this is good enough.

I didn't want to pass another argument through to every keyword, so this
moves `dynamic_scope` and `adjacent_results` into a `Context` object
that can be accessed as necessary. `short_circuit` is added there so
that it can be used in subschemas as well. The context object may also
be useful for extending things in the future.