Support `Schema#ref` in subschemas
`resolve_ref` needs to be called on the root schema because that's where
all the `resources` are stored.

The behavior here is kind of confusing because the ref is resolved just
like the `$ref` keyword would be in the schema, so it's dependent on the
schema's base URI. That means for a subschema `ref('#')` doesn't
necessarily resolve to itself (ie, if the subschema doesn't have `$id`).