Add Ruby 3.3 to test matrix
And update locked bundler to latest version that supports Ruby 2.5.

I had to add `base64`, `bigdecimal`, and `csv` to the gemspec because
Ruby 3.3 now warns that they won't be available in the standard library
in Ruby 3.4:

```
/Users/dharsha/repos/json_schemer/lib/json_schemer.rb:2: warning: base64 was loaded from the standard library, but will no longer be part of the default gems since Ruby 3.4.0. Add base64 to your Gemfile or gemspec.
/Users/dharsha/repos/json_schemer/lib/json_schemer.rb:3: warning: bigdecimal was loaded from the standard library, but will no longer be part of the default gems since Ruby 3.4.0. Add bigdecimal to your Gemfile or gemspec.
```

I'm not sure if they need version constraints or not, but I'm going to
leave it unset for now.

`minitest` is also triggering a warning, but it looks like it'll be
fixed in the next release: https://github.com/minitest/minitest/issues/969