Only parse `$schema` during schema validation
Parsing the whole schema can lead to errors for malformed keyword
values, eg:

```
>> JSONSchemer.schema({ 'properties' => '' })
/Users/dharsha/repos/json_schemer/lib/json_schemer/draft202012/vocab/applicator.rb:224:in `parse': undefined method `each_with_object' for an instance of String (NoMethodError)

            value.each_with_object({}) do |(property, subschema), out|
                 ^^^^^^^^^^^^^^^^^
	from /Users/dharsha/repos/json_schemer/lib/json_schemer/keyword.rb:14:in `initialize'
```

Instead, this creates a minimal parseable schema with just the `$schema`
value, if it's present and a string. That way the meta schema can be
determined as usual and then used to validate the provided schema.

Addresses: https://github.com/davishmcclurg/json_schemer/issues/167