Allow passing options to schema validation methods
The default meta schemas don't inherit the provided options, because
they're initialized with their own options and cached. To override
validation-time options, they need to be passed directly to the
validation methods.

This is a little brittle because the validation-time options are listed
separately to pull them out of the options hash. I couldn't think of a
better way to split the initialization and validation options.