Parse string meta schemas in `Schema` class
This allows string values in `Configuration#meta_schema` in order to be
consistent with `JSONSchemer.schema`.

There's a slight behavior change in that unknown `meta_schema` values
will be passed to the provided ref resolver. I think this is good
because it better matches the behavior of the `$schema` keyword.

`UnsupportedMetaSchema` is gone; `UnknownRef` will be raised instead.

It was necessary to add `$schema` in
`test_schema_validation_parse_error_with_custom_meta_schema` because
`ref_resolver` schemas inherit the provided `meta_schema`, which meant
the custom meta schema was using itself as a meta schema when provided
as a string.