Add `configuration` argument
This makes it easy to pass in an existing configuration object and
override multiple arguments without modifying the global configuration.

It also provides a simpler way to pass configuration to subschemas (and
ref schemas). `base_uri`, `meta_schema`, `ref_resolver`, and
`regexp_resolver` still need to be passed separately because they're
parsed and modified in the schema object (not reflected in the
configuration object).